test(extension): cover component extension install in lifecycle
