+++ src/core/extension/lifecycle.rs
@@ -779,9 +779,112 @@ pub fn read_source_revision(extension_id: &str) -> Option<String> {
 
 #[cfg(test)]
 mod tests {
-    use super::is_workdir_clean;
+    use super::{install, install_for_component, is_workdir_clean};
+    use crate::component;
+    use crate::test_support::with_isolated_home;
+    use std::fs;
+    use std::path::Path;
     use std::process::Command;
     use tempfile::TempDir;
+
+    fn write_extension_fixture(root: &Path, id: &str) {
+        let dir = root.join(id);
+        fs::create_dir_all(&dir).expect("extension dir");
+        fs::write(
+            dir.join(format!("{}.json", id)),
+            format!(
+                r#"{{
+  "name": "{} extension",
+  "version": "1.0.0"
+}}"#,
+                id
+            ),
+        )
+        .expect("extension manifest");
+    }
+
+    fn write_component_fixture(root: &Path, extensions: &[&str]) {
+        let extension_json = extensions
+            .iter()
+            .map(|id| format!(r#"    "{}": {{}}"#, id))
+            .collect::<Vec<_>>()
+            .join(",\n");
+
+        fs::write(
+            root.join("homeboy.json"),
+            format!(
+                r#"{{
+  "id": "multi-extension-component",
+  "extensions": {{
+{}
+  }}
+}}"#,
+                extension_json
+            ),
+        )
+        .expect("component config");
+    }
+
+    #[test]
+    fn test_install_for_component_installs_multiple_extensions() {
+        with_isolated_home(|home| {
+            let home = home.path();
+            let source = home.join("source");
+            write_extension_fixture(&source, "alpha");
+            write_extension_fixture(&source, "beta");
+
+            let component_dir = home.join("component");
+            fs::create_dir_all(&component_dir).expect("component dir");
+            write_component_fixture(&component_dir, &["alpha", "beta"]);
+            let component = component::discover_from_portable(&component_dir).expect("component");
+
+            let result = install_for_component(&component, &source.to_string_lossy())
+                .expect("install should succeed");
+
+            let installed_ids = result
+                .installed
+                .iter()
+                .map(|entry| entry.extension_id.as_str())
+                .collect::<Vec<_>>();
+            assert_eq!(installed_ids, vec!["alpha", "beta"]);
+            assert!(result.skipped.is_empty());
+            assert!(home
+                .join(".config/homeboy/extensions/alpha/alpha.json")
+                .exists());
+            assert!(home
+                .join(".config/homeboy/extensions/beta/beta.json")
+                .exists());
+        });
+    }
+
+    #[test]
+    fn test_install_for_component_skips_already_installed_extensions() {
+        with_isolated_home(|home| {
+            let home = home.path();
+            let source = home.join("source");
+            write_extension_fixture(&source, "alpha");
+            write_extension_fixture(&source, "beta");
+
+            let component_dir = home.join("component");
+            fs::create_dir_all(&component_dir).expect("component dir");
+            write_component_fixture(&component_dir, &["alpha", "beta"]);
+            let component = component::discover_from_portable(&component_dir).expect("component");
+
+            install(&source.join("alpha").to_string_lossy(), Some("alpha"))
+                .expect("pre-install alpha");
+
+            let result = install_for_component(&component, &source.to_string_lossy())
+                .expect("install should succeed");
+
+            let installed_ids = result
+                .installed
+                .iter()
+                .map(|entry| entry.extension_id.as_str())
+                .collect::<Vec<_>>();
+            assert_eq!(installed_ids, vec!["beta"]);
+            assert_eq!(result.skipped, vec!["alpha"]);
+        });
+    }
 
     #[test]
     fn is_workdir_clean_non_git_dir_returns_true() {
